More lighting fixtures and documentation (#9404)
* Fixes roogue  Light values

* Bright lasers

* converts and adds brighter light for lasers

* Some balancing

* Cleanup and documentation

+++ UnityProject/Assets/Scripts/Objects/Wallmounts/WallmountSpriteBehavior.cs
@@ -61,10 +61,10 @@ namespace Objects.Wallmounts
 				if (Camera2DFollow.followControl.target == null) return;
 				visible = wallmountBehavior.IsFacingPosition(Camera2DFollow.followControl.target.position);
 			}
-			SetAlpha(visible ? 1 : 0);
+			SetAlpha(visible ?1 : 0f);
 		}
 
-		protected virtual void SetAlpha(int alpha)
+		protected virtual void SetAlpha(float alpha)
 		{
 			spriteRenderer.color = new Color(spriteRenderer.color.r, spriteRenderer.color.g, spriteRenderer.color.b, alpha);
 		}
